fix(re-org-txs): addressed pr feedback
