--- src/tests-mempool/teardown.ts
@@ -1,5 +1,0 @@
-// ts-unused-exports:disable-next-line
-export default (): void => {
-  console.log('Jest - teardown');
-  console.log('Jest - teardown done');
-};
